refactor(infra): tighten docker package altitude
Two altitude fixes from the simplify review:

- Move the input-config write into the exchange module
  (exchange.write_config), which owns the exchange-dir lifecycle, instead
  of the facade reaching into that dir directly.
- Drop the facade re-exports of append_nccl_env / append_package_dispatch
  (shrinking __all__ to DockerRunner) and repoint the sole consumer,
  study.baseline_container, plus the two probe tests, at their real home
  infra.docker.command.
